requirements.txt
mem0ai
click>=8.0.0
rich>=13.0.0
python-dateutil>=2.8.0
pyyaml>=6.0
streamlit>=1.28.0
pandas>=2.0.0
python-dotenv>=1.0.0 